Fix deprecated non-canonical casts to canonical forms
Replace deprecated non-canonical PHP casts with their canonical equivalents:
- (integer) → (int)
- (boolean) → (bool)
- (double) / (real) → (float)
- (binary) → (string)

This addresses PHP 8.5 deprecation warnings for non-canonical cast syntax.

+++ framework/web/CHttpRequest.php
@@ -972,11 +972,11 @@ class CHttpRequest extends CApplicationComponent
 						if($matches[4][$i]==='q')
 						{
 							// sanity check on q value
-							$q=(double)$matches[5][$i];
+							$q=(float)$matches[5][$i];
 							if($q>1)
-								$q=(double)1;
+								$q=(float)1;
 							elseif($q<0)
-								$q=(double)0;
+								$q=(float)0;
 							$accept['params'][$matches[4][$i]]=$q;
 						}
 						else
@@ -987,7 +987,7 @@ class CHttpRequest extends CApplicationComponent
 				}
 				// q defaults to 1 if not explicitly given
 				if(!isset($accept['params']['q']))
-					$accept['params']['q']=(double)1;
+					$accept['params']['q']=(float)1;
 				$accepts[] = $accept;
 			}
 		}
